+++ src/Text/CommonMark/Sub.hs
@@ -1,5 +1,10 @@
+{-# LANGUAGE DisambiguateRecordFields #-}
+{-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE NamedFieldPuns #-}
 module Text.CommonMark.Sub
-    ( Level
+    ( HeadingPattern (..)
+    , HeadingTitlePattern (..)
+    , Level
     , extractSection
     , flattenInlineNodes
     , matchesHeading
@@ -8,26 +13,71 @@ module Text.CommonMark.Sub
 import Prelude hiding (concat)
 
 import CMark
-import Data.Text (Text, concat, pack, strip)
+import Data.Text (Text, concat, pack, strip, toLower)
+import Text.Regex.TDFA
+import Text.Regex.TDFA.Text
 
-extractSection :: Level -> (Text -> Text -> Bool) -> Text -> Node -> Node
-extractSection headingLevel equals headingTitle (Node pos DOCUMENT nodes) =
+data HeadingPattern = HeadingPattern
+    { headingLevel :: Level
+    , caseInsensitive :: Bool
+    , titlePattern :: HeadingTitlePattern
+    } deriving (Eq, Ord, Show)
+
+data HeadingTitlePattern
+    = HeadingTitleRegex Text
+    | HeadingTitleText Text
+    deriving (Eq, Ord, Show)
+
+extractSection :: HeadingPattern -> Node -> Node
+extractSection pat@HeadingPattern { headingLevel } (Node pos DOCUMENT nodes) =
     Node pos DOCUMENT $ case headlessNodes of
-        (x : xs) -> x : takeWhile (isInSection headingLevel) xs
+        (x : xs) -> x : takeWhile isInSection xs
         [] -> []
   where
-    isInSection :: Level -> Node -> Bool
-    isInSection level (Node _ (HEADING lv) _) = lv > level
-    isInSection _ _ = True
+    isInSection :: Node -> Bool
+    isInSection (Node _ (HEADING lv) _) = lv > headingLevel
+    isInSection _ = True
     headlessNodes :: [Node]
     headlessNodes =
-        dropWhile (not . matchesHeading headingLevel equals headingTitle) nodes
-extractSection _ _ _ (Node pos _ _) = Node pos DOCUMENT []
+        dropWhile (not . matchesHeading pat) nodes
+extractSection _ (Node pos _ _) = Node pos DOCUMENT []
 
-matchesHeading :: Level -> (Text -> Text -> Bool) -> Text -> Node -> Bool
-matchesHeading level equals text (Node _ (HEADING lv) nodes) =
-    lv == level && flattenInlineNodes nodes `equals` text
-matchesHeading _ _ _ _ = False
+matchesHeading :: HeadingPattern -> Node -> Bool
+matchesHeading HeadingPattern { headingLevel
+                              , caseInsensitive
+                              , titlePattern
+                              } = case titlePattern of
+    HeadingTitleText text -> if caseInsensitive
+        then \ case
+            (Node _ (HEADING lv) nodes) ->
+                lv == headingLevel &&
+                    toLower (flattenInlineNodes nodes) == toLower text
+            _ -> False
+        else \ case
+            (Node _ (HEADING lv) nodes) ->
+                lv == headingLevel && flattenInlineNodes nodes == text
+            _ -> False
+    HeadingTitleRegex regexPat ->
+        let compOption = CompOption
+                { caseSensitive = not caseInsensitive
+                , multiline = False
+                , rightAssoc = True
+                , newSyntax = True
+                , lastStarGreedy = True
+                }
+            execOption = ExecOption { captureGroups = False }
+            regex' = compile compOption execOption regexPat
+        in
+            case regex' of
+                Left _ -> const False
+                Right regex ->
+                    \ case
+                    (Node _ (HEADING lv) nodes) ->
+                        lv == headingLevel &&
+                            case execute regex (flattenInlineNodes nodes) of
+                                Right (Just _) -> True
+                                _ -> False
+                    _ -> False
 
 flattenInlineNodes :: [Node] -> Text
 flattenInlineNodes =